feat: Improve project structure and UI responsiveness
- Added `.cursor/` to `.gitignore` for Cursor IDE settings.
- Introduced a new target in `IndigoMovieManager.csproj` to copy development tools for local releases.
- Enhanced `MainWindow.xaml` by replacing the `Image` control with a `MediaElement` for better media handling and added a fallback `Image`.
- Updated `MainWindow.xaml.cs` to implement caching for filtered movie records, improving performance during searches.
- Refactored `PlayPositionResolver` to streamline play position calculations based on image clicks, enhancing user interaction.

+++ Tools.cs
@@ -181,7 +181,7 @@ namespace IndigoMovieManager
                                         }
 
                                         //1つ目
-                                        long sec = BitConverter.ToUInt16(bufInf, 0);
+                                        long sec = BitConverter.ToInt32(bufInf, 0);
                                         int i = 0;
                                         thumbSec[i] = (int)sec;
                                         i++;
@@ -194,7 +194,7 @@ namespace IndigoMovieManager
                                             {
                                                 break;
                                             }
-                                            sec = BitConverter.ToUInt16(bufInf, 0);
+                                            sec = BitConverter.ToInt32(bufInf, 0);
                                             thumbSec[i] = (int)sec;
                                             i++;
                                         }
